Updated Common version.

+++ data/scripts/upgrade.php
@@ -23,6 +23,14 @@ $connection = $services->get('Omeka\Connection');
 $messenger = $plugins->get('messenger');
 $entityManager = $services->get('Omeka\EntityManager');
 
+if (!method_exists($this, 'checkModuleActiveVersion') || !$this->checkModuleActiveVersion('Common', '3.4.57')) {
+    $message = new \Omeka\Stdlib\Message(
+        $translate('The module %1$s should be upgraded to version %2$s or later.'), // @translate
+        'Common', '3.4.57'
+    );
+    throw new \Omeka\Module\Exception\ModuleCannotInstallException((string) $message);
+}
+
 if (version_compare($oldVersion, '3.3.4.2', '<')) {
     $settings->set('statistics_public_allow_browse', $settings->get('statistics_public_allow_browse_pages', false));
     $settings->delete('statistics_public_allow_browse_pages');
@@ -220,12 +228,4 @@ WHERE `site_id` = 0
 ;
 SQL;
     $connection->executeStatement($sql);
-
-    if (!method_exists($this, 'checkModuleActiveVersion') || !$this->checkModuleActiveVersion('Common', '3.4.51')) {
-        $message = new Message(
-            'The module %1$s should be upgraded to version %2$s or later.', // @translate
-            'Common', '3.4.51'
-        );
-       throw new \Omeka\Module\Exception\ModuleCannotInstallException((string) $message);
-    }
 }
